Remove "non-myopic rewards are based on stake" test
It turns out that the total stake of the test cluster used for integration is so large that the given user stakes only make a minuscule difference. Our new non-myopic reward calculation uses `Double` internally for reasons of speed, which causes this test fail due to limited precision.

In order to create larger results in this test, one would have to consider a local cluster where the number of stake pools is larger than the number of desired pools (`nOpt` parameter) and apply the test to a pool with low desirability and low owner stake, so that it will drop out of the ranking and the user stake makes all the difference to the rewards. (I have done such a test manually.)

However, as non-myopic rewards are going to be phased out in the future, I believe that the appropriate solution is to simply delete this test.